Move search logic to the stucture module
